Remove header, list and footer composer's.
