+++ QuizMobileApp/QuizApp/Composers/QuizComposer.swift
@@ -13,11 +13,13 @@ import QuiziOS
 public final class QuizComposer {
     private let quizQuestionLoader: QuestionLoader
     private var quizGameEngine: QuizGameEngine?
-    private var counter = QuizGameTimer(withSeconds: 300)
-    private var headerComposer = QuizHeaderComposer()
-    private var listComposer = QuizAnswerListComposer()
-    private var footerComposer = QuizFooterComposer()
+    private var counter = QuizGameTimer(withSeconds: 300)    
     private let messageComposer = QuizMessageComposer()
+    private var isPlaying = false
+
+    var headerPresenter: QuizHeaderPresenter?
+    var listPresenter: QuizAnswerListPresenter?
+    var footerPresenter: QuizFooterPresenter?
     
     public init(questionLoader: QuestionLoader) {
         self.quizQuestionLoader = MainQueueDispatchDecorator(decoratee: questionLoader)
@@ -25,9 +27,9 @@ public final class QuizComposer {
     
     public func quizRootViewController() -> QuizRootViewController {
         let controller = makeQuizViewController()
-        controller.quizHeaderController = headerComposer.header()
-        controller.quizAnswerListController = listComposer.answerListController()
-        controller.quizFooterController = footerComposer.footer()
+        controller.quizHeaderController = header()
+        controller.quizAnswerListController = answerListController()
+        controller.quizFooterController = footer()
         
         return controller
     }
@@ -45,10 +47,71 @@ public final class QuizComposer {
         let quizGameEngine = QuizGameEngine(counter: self.counter, correctAnswers: questionItem.answer)
         self.quizGameEngine = quizGameEngine
         self.quizGameEngine?.delegate = WeakRefVirtualProxy(self)
-        self.headerComposer.quizGameEngine = quizGameEngine
-        self.footerComposer.quizGameEngine = quizGameEngine
-        
         self.counter.delegate = WeakRefVirtualProxy(quizGameEngine)
+    }
+
+    func header() -> QuizHeaderViewController {
+        let controller = makeQuizHeaderViewController()
+        headerPresenter = QuizHeaderPresenter(quizHeader: WeakRefVirtualProxy(controller))
+
+        return controller
+    }
+
+    private func makeQuizHeaderViewController() -> QuizHeaderViewController {
+        let bundle = Bundle(for: QuizRootViewController.self)
+        let headerStoryboard = UIStoryboard(name: "QuizHeader", bundle: bundle)
+        let quizHeaderController = headerStoryboard.instantiateInitialViewController() as! QuizHeaderViewController
+        quizHeaderController.delegate = self
+
+        return quizHeaderController
+    }
+
+    func answerListController() -> QuizAnswerListViewController {
+        let controller = makeQuizAnswerListViewController()
+        listPresenter = QuizAnswerListPresenter(answerList: controller)
+        return controller
+    }
+
+    private func makeQuizAnswerListViewController() -> QuizAnswerListViewController {
+        let bundle = Bundle(for: QuizAnswerListViewController.self)
+        let answerListStoryboard = UIStoryboard(name: "QuizAnswerList", bundle: bundle)
+        let quizAnswerListController = answerListStoryboard.instantiateInitialViewController() as! QuizAnswerListViewController
+
+        return quizAnswerListController
+    }
+
+    func footer() -> QuizFooterViewController {
+        let controller = makeQuizFooterViewController()
+        footerPresenter = QuizFooterPresenter(quizFooter: WeakRefVirtualProxy(controller))
+
+        return controller
+    }
+
+    private func makeQuizFooterViewController() -> QuizFooterViewController {
+        let bundle = Bundle(for: QuizFooterViewController.self)
+        let footerStoryboard = UIStoryboard(name: "QuizFooter", bundle: bundle)
+        let quizFooterController = footerStoryboard.instantiateInitialViewController() as! QuizFooterViewController
+        quizFooterController.delegate = self
+
+        return quizFooterController
+    }
+}
+
+extension QuizComposer: QuizHeaderViewControllerDelegate {
+    public func didTapNewAnswer(_ answer: String) {
+        quizGameEngine?.addAnswer(answer)
+    }
+}
+
+extension QuizComposer: QuizFooterViewControllerDelegate {
+    public func didClickStatusButton() {
+        if isPlaying {
+            isPlaying = false
+            quizGameEngine?.reset()
+        } else {
+            isPlaying = true
+            quizGameEngine?.start()
+        }
     }
 }
 
@@ -60,9 +123,10 @@ extension QuizComposer: QuizRootViewControllerDelegate {
             switch result {
             case let .success(questions):
                 guard let questionItem = questions.first else { return }
-                                
-                self?.headerComposer.headerPresenter?.didFinishLoadGame(with: questionItem)
-                self?.footerComposer.presenter?.didFinishLoadGame(with: questionItem)
+
+                self?.headerPresenter?.didFinishLoadGame(with: questionItem)
+
+                self?.footerPresenter?.didFinishLoadGame(with: questionItem)
                 
                 self?.startGameEngine(questionItem)
                 
@@ -81,9 +145,9 @@ extension QuizComposer: QuizGameDelegate {
         if gameStatus.isGameFinished {
             messageComposer.showFinishedGame(gameStatus)
         } else {
-            headerComposer.headerPresenter?.didUpdateGameStatus(gameStatus)
-            listComposer.updateList(gameStatus)
-            footerComposer.presenter?.didUpdateGameStatus(gameStatus)
+            headerPresenter?.didUpdateGameStatus(gameStatus)
+            listPresenter?.didUpdateGameStatus(gameStatus)
+            footerPresenter?.didUpdateGameStatus(gameStatus)
         }
     }
 }
